Result analyzer

+++ util.sh
@@ -9,13 +9,11 @@ for job in $(qstat -u iferreira | awk 'NR>5 {print $1}'); do
     fi
 done
 
-echo
-
-while IFS= read -r dir; do
-    if [[ -z "${jobnames[$dir]}" ]]; then
+find experiments -type d | while read -r dir; do
+    if [[ -z "$(find "$dir" -mindepth 1 -type d 2>/dev/null)" && \
+          ! -f "$dir/accuracy.png" && \
+          -z "${jobnames[$dir]}" ]]; then
         echo "Deleting $dir"
         rm -rf "$dir"
     fi
-done < <(find experiments -type d -empty)
-echo
-
+done
